requirements.txt
fastapi
uvicorn
supabase
google-generativeai
websockets